Add pri.raise before cond branch except loop exit
Now the pass will insert pri.raise before all cond branch unless that
branch is the loop exit conditional branch.
For the loop, consider following mir:
loop:
    ...
    BLT x5, x6, loop;
We shouldn't insert  pri raise right before BLT, for that would cause
priority saturation if the loop iteration for more than 63 times.

To deal with loop, we insert fsa.pri.raise inst. at IDom of the loop BB,
where we can utilize MachineLoopInfo class to identify whether a cond
branch is inside a loop or not.
If the Idom of the cond branch is not found, skip that cond branch.