Set priority based on PDom level
In the LLVM PDom tree, nodes that dominate more BBs (basic blocks) play
a more critical role in program execution. We aim to increase the
chances of different threads reconverging during execution. Therefore,
when a thread reaches a node that dominates many BBs, that BB should
have lower execution priority. This allows other threads to also reach
this BB, enhancing the likelihood of reconvergence.

Typically, when a BB dominates more nodes, it is at a shallower level
in the PDom tree (as it is closer to the virtual root, which dominates
all nodes). Therefore, we can adjust priorities based on a BB's depth
in the PDom tree: the shallower the BB, the lower its priority.
Conversely, BBs with deeper level in PDom tree usually dominate fewer
BBs and have less chance to reconverge, so their priority should be
higher to quickly skip such BB and reach reconvergence opportunities
(BBs with shallower level).

In our priority adjustment instructions, smaller numbers indicate
lower priority, and larger numbers indicate higher priority. Thus, we
can directly use a BB's depth in the PDom tree as its priority.