Only skip insertion after first insertion of inst
The previous process of insertion skipping check is as follow:
1. Mark current BB's insertion to be skippable.
2. Check if all BB's predecessors's CPNLevel is as same as current BB,
if not, mark current BB's insertion to be unskippable.
3. Based on result of 2. to determine whether skip insertion.

However, above flow cannot apply on entry block which has no
predecessors and cause such BB's insertion to be skipped. To address
such problem, we can do following:

Skip insertion of the `pri.set` instruction ONLY if there has already
been a inserted instruction previously (indicated by MadeChange being
true). This is especially for the entry block, which has no predecessors
and therefore does not need a redundant check. To ensure that a `pri.set`
instruction is inserted in the entry block, we initialize
"IsRedundantInsertion" to false (the value of MadeChange) at the beginning
of processing for each block.

Afterall, new process become:
1. Init MadeChange to be false
2. Mark current BB's to be (MadeChange) ? skippable : unskippable
3. Check if all BB's predecessors's CPNLevel is as same as current BB,
if not, mark current BB's insertion to be unskippable.
4. Based on result of 3. to determine whether skip insertion.
5. If an insertion is completed, set MadeChange to be true.