Add frontend flags for ICS-first alg
Currently, the ICS-first algorithm, which operates at the frontend
level, is only implemented in the backend through specific flags. As a
result, PoCL is unable to take advantage of ICS-first due to its
two-stage compilation flow:
    Stage 1: Source code → LLVM IR (frontend)
    Stage 2: LLVM IR → LLVM MIR → Target object
By the time PoCL reaches Stage 2, it is already too late to utilize the
ICS flag, since ICS-first is designed to work in the frontend.

In this commit, a code gen flag `fsa-ics-first-cg` is introduced as a
frontend flag for ICS-first alg.